v1.2: Buttons Shortcode & Minor Improvments

+++ app-store-redirect.php
@@ -3,7 +3,7 @@
 /**
  * Plugin Name: App Store Redirect
  * Description: Redirects users to the appropriate app store based on their operating system.
- * Version: 1.1
+ * Version: 1.2
  * Author: Somar Kesen
  * Author URI: https://github.com/somarkn99
  * Text Domain: app-store-redirect
@@ -12,6 +12,9 @@
 define('ANDROID_APP_URL_OPTION', 'android_app_url');
 define('IOS_APP_URL_OPTION', 'ios_app_url');
 define('CUSTOM_ROUTE_OPTION', 'custom_route');
+
+// ------------------------------------------------------------------------------------------------
+// Init & Admin dashboard Settings
 
 // Load Translation Files
 function app_store_redirect_load_textdomain()
@@ -52,6 +55,13 @@ function app_store_redirect_settings_page()
         // Save the Android and iOS app URLs and the custom route
         update_option(ANDROID_APP_URL_OPTION, esc_url($_POST['android_app_url']));
         update_option(IOS_APP_URL_OPTION, esc_url($_POST['ios_app_url']));
+
+        // Check if the custom route is already in use
+        $existing_pages = get_pages(['meta_key' => '_wp_page_template', 'meta_value' => 'page-templates/template-custom.php']);
+        if (!empty($existing_pages)) {
+            echo '<div class="error"><p>' . __('Custom route is already in use. Please choose a different route.', 'app-store-redirect') . '</p></div>';
+            return;
+        }
         update_option(CUSTOM_ROUTE_OPTION, esc_html($_POST['custom_route']));
 
         // Flush old Cached rewrite rules without visiting Settings->Permalinks (effective on the next website load)
@@ -68,11 +78,11 @@ function app_store_redirect_settings_page()
         <form method="post" action="">
             <?php wp_nonce_field('app_store_redirect_update', 'app_store_redirect_nonce'); ?>
             <label for="android_app_url"><?= __('Google Play URL:', 'app-store-redirect') ?></label>
-            <input type="text" name="android_app_url" placeholder="<?= __('Your App link on Google Play', 'app-store-redirect') ?>" id="android_app_url" value="<?php echo esc_attr($android_app_url); ?>" size="60">
+            <input type="url" name="android_app_url" placeholder="<?= __('Your App link on Google Play', 'app-store-redirect') ?>" id="android_app_url" value="<?php echo esc_attr($android_app_url); ?>" size="60">
             <br>
             <br>
             <label for="ios_app_url"><?= __('Apple Store URL:', 'app-store-redirect') ?></label>
-            <input type="text" name="ios_app_url" placeholder="<?= __('Your App link on Apple Store', 'app-store-redirect') ?>" id="ios_app_url" value="<?php echo esc_attr($ios_app_url); ?>" size="60">
+            <input type="url" name="ios_app_url" placeholder="<?= __('Your App link on Apple Store', 'app-store-redirect') ?>" id="ios_app_url" value="<?php echo esc_attr($ios_app_url); ?>" size="60">
             <br>
             <br>
             <label for="custom_route"><?= __('Custom Route:', 'app-store-redirect') ?></label>
@@ -82,8 +92,118 @@ function app_store_redirect_settings_page()
             <input type="submit" name="app_store_redirect_submit" class="button button-primary" value="<?= __('Save') ?>">
         </form>
     </div>
+    <div class="wrap-preview-shortcode">
+        <h3><?=__('Available Shortcodes', 'app-store-redirect')?></h3>
+        <span><?=__('(click to copy)', 'app-store-redirect')?></span>
+
+        <h4>[app-store-button buttons="ios"]</h4>
+        <?= do_shortcode('[app-store-button buttons="ios"]');?>
+
+        <h4>[app-store-button buttons="android"]</h4>
+        <?= do_shortcode('[app-store-button buttons="android"]');?>
+
+        <h4>[app-store-button buttons="all"]</h4>
+        <?= do_shortcode('[app-store-button buttons="all"]');?>
+
+        <h4>[app-store-button buttons="auto"]</h4>
+        <?= do_shortcode('[app-store-button buttons="ios"]');?>
+        <span style="display: inline;"><?=__('or', 'app-store-redirect')?></span>
+        <?= do_shortcode('[app-store-button buttons="android"]');?>
+        <span style="display: inline;"><?=__('According to the user operating system', 'app-store-redirect')?></span>
+    </div>
+
+    <style>
+        .wrap-preview-shortcode h4{
+            cursor: pointer;
+        }
+    </style>
+    <!-- Copy on Click -->
+    <script>
+        document.querySelectorAll('.wrap-preview-shortcode h4').forEach(function(h4) {
+            h4.addEventListener('click', function() {
+                var text = this.textContent;
+                var textarea = document.createElement('textarea');
+                textarea.textContent = text;
+                document.body.appendChild(textarea);
+                textarea.select();
+                document.execCommand('copy');
+                document.body.removeChild(textarea);
+                alert("<?=__('Shortcode Copied !', 'app-store-redirect')?>");
+            });
+        });
+
+    </script>
 <?php
 }
+
+
+// ------------------------------------------------------------------------------------------------
+// Shortcode
+
+/*
+ * Shortcode: app-store-button 
+ * Usability: [app-store-button buttons="auto"]
+ * 
+ * Attribuites:
+ * buttons
+ *   - auto : Show button according to the user operatoin-system
+ *   - all : (default) Show all buttons
+ *   - android : Show only google play
+ *   - ios : Show only app store
+ * 
+ */
+add_action( 'init', 'app_store_redirect_register_button_shortcode');
+function app_store_redirect_register_button_shortcode(){
+    add_shortcode('app-store-button' , 'app_store_redirect_button_shortcode');
+}
+function app_store_redirect_button_shortcode($atts){
+
+    wp_enqueue_style( 'app-store-button-style', plugins_url( '/style.css', __FILE__ ), array(), '1.0', 'all' );
+    extract(shortcode_atts([
+        'buttons' => 'all',
+    ], $atts));
+
+    // All available buttons
+    $ARGS = [
+        'ios' => [
+            'button_title'     => __('App Store','app-store-redirect'),
+            'button_sub_title' => __('Download on the','app-store-redirect'),
+            'button_class'     => 'apple-btn',
+            'app_link_setting' => IOS_APP_URL_OPTION
+        ],
+        'android' => [
+            'button_title'     => __('Google Play','app-store-redirect'),
+            'button_sub_title' => __('Get it on','app-store-redirect'),
+            'button_class'     => 'google-btn',
+            'app_link_setting' => ANDROID_APP_URL_OPTION
+        ]
+    ];
+
+    $user_os = app_store_redirect_get_user_os();
+
+    // Echoing html directly causes errors with Gutenberg editor.
+    ob_start();
+
+    if($buttons == 'auto' && array_key_exists($user_os, $ARGS)){ // if desktop, default to all buttons
+        $data = $ARGS[$user_os];
+        include __DIR__ . '/shortcode-app-store-button.php';
+
+    }
+    elseif(array_key_exists($buttons, $ARGS)){
+        $data = $ARGS[$buttons];
+        include __DIR__ . '/shortcode-app-store-button.php';
+    }
+    else{ //  all or unknown user agent or settings page preview
+        foreach($ARGS as $data){
+            include __DIR__ . '/shortcode-app-store-button.php';
+        }
+    }
+    return ob_get_clean();
+}
+
+
+// ------------------------------------------------------------------------------------------------
+// Redirection Handling
 
 // Add a custom route based on the setting
 function app_store_redirect_init()
@@ -91,14 +211,6 @@ function app_store_redirect_init()
     $custom_route = get_option(CUSTOM_ROUTE_OPTION);
 
     if (!empty($custom_route)) {
-        // Check if the custom route is already in use
-        $existing_pages = get_pages(array('meta_key' => '_wp_page_template', 'meta_value' => 'page-templates/template-custom.php'));
-
-        if (!empty($existing_pages)) {
-            echo '<div class="error"><p>' . __('Custom route is already in use. Please choose a different route.', 'app-store-redirect') . '</p></div>';
-            return;
-        }
-
         add_rewrite_rule('^' . $custom_route . '/?$', 'index.php?app_store_redirect=true', 'top');
     }
 }
@@ -115,19 +227,17 @@ add_filter('query_vars', 'app_store_redirect_query_vars');
 // Redirect users to the appropriate app store based on settings
 function app_store_redirect_template_redirect()
 {
-    global $wp_query;
-
     if (get_query_var('app_store_redirect')) {
-        $user_agent = $_SERVER['HTTP_USER_AGENT'];
-
+        
+        $user_os = app_store_redirect_get_user_os();
         $android_app_url = get_option(ANDROID_APP_URL_OPTION, '');
         $ios_app_url = get_option(IOS_APP_URL_OPTION, '');
 
-        if (strpos($user_agent, 'Android') !== false && !empty($android_app_url)) {
-            header("Location: $android_app_url");
+        if ($user_os == 'android' && !empty($android_app_url)) {
+            wp_redirect("Location: $android_app_url");
             exit;
-        } elseif ((strpos($user_agent, 'iPhone') !== false || strpos($user_agent, 'iPad') !== false) && !empty($ios_app_url)) {
-            header("Location: $ios_app_url");
+        } elseif ($user_os == 'ios' && !empty($ios_app_url)) {
+            wp_redirect("Location: $ios_app_url");
             exit;
         } else {
             wp_safe_redirect(home_url());
@@ -136,3 +246,15 @@ function app_store_redirect_template_redirect()
     }
 }
 add_action('template_redirect', 'app_store_redirect_template_redirect');
+
+
+function app_store_redirect_get_user_os(){
+    $user_agent = $_SERVER['HTTP_USER_AGENT'];
+    if (strpos($user_agent, 'Android') !== false) {
+        return 'android';
+    } elseif ((strpos($user_agent, 'iPhone') !== false || strpos($user_agent, 'iPad') !== false)) {
+        return 'ios';
+    } else {
+        return 'desktop';
+    }
+}
